+++ MoltenVK/MoltenVK/API/vk_mvk_moltenvk.h
@@ -322,22 +322,23 @@ typedef struct {
 	VkBool32 presentWithCommandBuffer;
 
 	/**
-	 * If enabled, swapchain images will use simple Nearest sampling when magnifying the
-	 * swapchain image to fit a physical display surface. If disabled, swapchain images will
+	 * If enabled, swapchain images will use simple Nearest sampling when minifying or magnifying
+	 * the swapchain image to fit a physical display surface. If disabled, swapchain images will
 	 * use Linear sampling when magnifying the swapchain image to fit a physical display surface.
 	 * Enabling this setting avoids smearing effects when swapchain images are simple interger
 	 * multiples of display pixels (eg- macOS Retina, and typical of graphics apps and games),
 	 * but may cause aliasing effects when using non-integer display scaling.
 	 *
-	 * The value of this parameter may be changed before creating a VkSwapchain,
+	 * The value of this parameter must be changed before creating a VkSwapchain,
 	 * for the change to take effect.
 	 *
 	 * The initial value or this parameter is set by the
-	 * MVK_CONFIG_SWAPCHAIN_MAG_FILTER_USE_NEAREST
+	 * MVK_CONFIG_SWAPCHAIN_MIN_MAG_FILTER_USE_NEAREST
 	 * runtime environment variable or MoltenVK compile-time build setting.
 	 * If neither is set, the value of this parameter defaults to true.
 	 */
-	VkBool32 swapchainMagFilterUseNearest;
+	VkBool32 swapchainMinMagFilterUseNearest;
+#define swapchainMagFilterUseNearest swapchainMinMagFilterUseNearest
 
 	/**
 	 * The maximum amount of time, in nanoseconds, to wait for a Metal library, function, or